extract generic session helpers from local
