+++ modules/common/sessions/index.ts
@@ -1,2 +1,5 @@
 export * from "./getCurrentPlr";
 export * from "./isWaitingForRemote";
+export * from "./newSessionId";
+export * from "./session2battle";
+export * from "./updateSession";
