+++ modules/local/src/session/index.ts
@@ -1,6 +1,3 @@
 export * from "./newSessionFromBattle";
-export * from "./updateSession";
-export * from "./session2battle";
-export * from "./newSessionId";
 export * from "./forkSessionFromBattle";
 export * from "./importSessionFromBattle";
